% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_kgram_freqs.R
\name{as_kgram_freqs}
\alias{as_kgram_freqs}
\title{Coerce to k-gram frequency table}
\usage{
as_kgram_freqs(object, ...)
}
\arguments{
\item{object}{object to be coerced to \code{kgram_freqs}.}

\item{...}{further argument passed to or from other methods.}
}
\value{
a \code{kgram_freqs} object.
}
\description{
At present, this is just an utility for recovering a \code{kgram_freqs}
object from a \link[kgrams]{language_model}.
}
\examples{
model <- language_model( kgram_freqs("a b b a b", 3) )
freqs <- as_kgram_freqs(model)
}
\author{
Valerio Gherardi
}
