% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_op.R
\name{conditional}
\alias{conditional}
\alias{\%|\%}
\title{Word-context conditional expression}
\usage{
word \%|\% context
}
\description{
Create word-context conditional expression with the infix notation.
}
