% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_concatenation.R
\name{\%+\%}
\alias{\%+\%}
\title{String concatenation}
\usage{
lhs \%+\% rhs
}
\arguments{
\item{lhs}{a string or vector of strings.}

\item{rhs}{a string or vector of strings.}
}
\value{
a string or vector of strings.
}
\description{
String concatenation
}
\details{
The expression \code{lhs \%+\% rhs} is equivalent to
\code{paste(lhs, rhs, sep = " ", collapse = NULL, recycle0 = FALSE)}.
See \link[base]{paste} for more details.
}
\examples{
"i love" \%+\% c("cats", "jazz", "you")
}
\seealso{
\link[base]{paste}
}
\author{
Valerio Gherardi

Brief synthax for string concatenation.
}
