% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midsummer}
\alias{midsummer}
\title{A Midsummer Night's Dream}
\format{
A length one character vector, containing the entire text of
"A Midsummer Night's Dream" from William Shakespeare. The script used
for generating this file is available
\href{https://github.com/vgherard/kgrams/blob/main/data-raw/shakespeare.R}{here}
}
\source{
\url{https://www.folger.edu/}
}
\usage{
midsummer
}
\description{
The entire play "A Midsummer Night's Dream" from William Shakespeare.
}
\examples{
midsummer[840]

}
\seealso{
\link[kgrams]{much_ado}
}
\keyword{datasets}
