% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{much_ado}
\alias{much_ado}
\title{Much Ado About Nothing}
\format{
A length one character vector, containing the entire text of
"Much Ado About Nothing" from William Shakespeare. The script used
for generating this file is available
\href{https://github.com/vgherard/kgrams/blob/main/data-raw/shakespeare.R}{here}
}
\source{
\url{http://shakespeare.mit.edu/much_ado/}
}
\usage{
much_ado
}
\description{
The entire play "Much Ado About Nothing" from William Shakespeare.
}
\examples{
much_ado[840]

}
\seealso{
\link[kgrams]{midsummer}
}
\keyword{datasets}
