% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{smoothers}
\alias{smoothers}
\alias{info}
\title{k-gram Probability Smoothers}
\usage{
smoothers()

info(smoother)
}
\arguments{
\item{code}{a string. Code name of probability smoother.}
}
\value{
\code{smoothers()} returns a character vector, the list of code names
of probability smoothers available in \link[kgrams]{kgrams}.
\code{smoother_info(code)} returns \code{NULL} and prints some information on
the selected smoothing technique.
}
\description{
Informations on available k-gram continuation probability smoothers
}
\examples{
# List available smoothers
smoothers()

# Get information on smoother "kn", i.e. Interpolated Kneser-Ney
smoother_info("kn")
}
\author{
Valerio Gherardi
}
