% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgrams-package.R
\docType{package}
\name{kgrams-package}
\alias{kgrams}
\alias{kgrams-package}
\title{kgrams: k-grams and k-gram Language Models in R}
\description{
Build k-gram language models efficiently with an intuitive interface.
}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
