% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgrams-package.R
\docType{package}
\name{kgrams-package}
\alias{kgrams}
\alias{kgrams-package}
\title{kgrams: k-grams, Language Models, and All That}
\description{
Efficient tools for training and tuning k-gram language models, including several probability smoothing methods, perplexity computations, random text generation and more.
}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
